+++ discopy/drawing/drawing.py
@@ -572,20 +572,11 @@ class Drawing(TracedCategory, RichDisplay):
 
     @classmethod
     def permutation(cls, xs: Sequence[int], doms) -> Drawing:
-        """ Draw a permutation of the wires in ``dom``. """
+        """ Draw a permutation of the wires in ``doms``. """
         from discopy.symmetric import Permutation
-        xs = finset.Permutation(xs)
         dom = cls.ob().tensor(*doms)
+        xs = finset.Permutation(xs, len(dom))
         if xs.is_identity:
-            return Drawing.id(dom)
-        return Permutation(dom, xs).to_drawing()
-
-    @staticmethod
-    def permutation(xs, dom) -> Drawing:
-        """ Draw a permutation of the wires in ``dom``. """
-        from discopy.symmetric import Permutation
-        xs = list(xs)
-        if xs == list(range(len(xs))):
             return Drawing.id(dom)
         return Permutation(dom, xs).to_drawing()