+++ discopy/para.py
@@ -19,6 +19,7 @@ Summary
     Symmetric
     Traced
     Markov
+    Comarkov
     Closed
     Feedback
     Compact
@@ -104,10 +105,12 @@ from __future__ import annotations
 
 from dataclasses import dataclass
 
-from discopy import symmetric, markov, closed, feedback, compact, frobenius
+from discopy import (
+    symmetric, markov, comarkov, closed, feedback, compact, frobenius)
 from discopy.abc import (
-    ClosedCategory, CompactCategory, FeedbackCategory, HypergraphCategory,
-    MarkovCategory, NamedGeneric, SymmetricCategory, TracedCategory)
+    ClosedCategory, CompactCategory, ComarkovCategory, FeedbackCategory,
+    HypergraphCategory, MarkovCategory, NamedGeneric, SymmetricCategory,
+    TracedCategory)
 from discopy.utils import (
     AxiomError, assert_iscomposable, assert_isinstance, classproperty,
     unbiased)
@@ -283,6 +286,25 @@ class Markov(Symmetric, MarkovCategory):
         return cls.lift(cls.category.copy(x, n))
 
 
+class Comarkov(Symmetric, ComarkovCategory):
+    """
+    Parametric maps over a comarkov underlying `category` form a comarkov
+    category, with the merge of the underlying category as :meth:`merge`.
+    """
+    category = comarkov.Diagram
+
+    @classmethod
+    def merge(cls, x: Symmetric.ob, n: int = 2) -> Comarkov:
+        """
+        The merge of the underlying category, with empty parameter space.
+
+        Parameters:
+            x : The object to merge.
+            n : The number of copies.
+        """
+        return cls.lift(cls.category.merge(x, n))
+
+
 class Closed(Markov, ClosedCategory):
     """
     Parametric maps over a closed underlying `category` form a closed
@@ -322,7 +344,7 @@ class Closed(Markov, ClosedCategory):
         return type(self)(self.dom[:-n], inside.cod, self.param, inside)
 
 
-class Feedback(Markov, FeedbackCategory):
+class Feedback(Markov, Comarkov, FeedbackCategory):
     """
     Parametric maps over a feedback underlying `category` form a feedback
     category, with :meth:`delay` applied to all four components.
@@ -393,7 +415,7 @@ class Compact(Traced, CompactCategory):
     curry = Closed.curry
 
 
-class Hypergraph(Compact, Markov, HypergraphCategory):
+class Hypergraph(Compact, Markov, Comarkov, HypergraphCategory):
     """
     Parametric maps over a hypergraph underlying `category` form a
     hypergraph category, with the spiders of the underlying category.